fix(conductor): corregge SQL type errors in recordTaskFinish
Due bug in bot-metrics.ts recordTaskFinish che causavano:

1. task_id = $8 (text) vs colonna BIGINT → "operator is not unique:
   unknown - unknown". Fix: WHERE task_id = $8::bigint.
   Questo errore era salvato come task error_message in DB, facendo
   fallire task completati con successo (falso negativo).

2. $9 usato sia come BIGINT (ui_duration_ms colonna) sia come
   DOUBLE PRECISION (nella somma con EXTRACT() che ritorna float) →
   "inconsistent types deduced for parameter $9". Fix: cast espliciti
   $9::bigint e EXTRACT(...)::bigint * 1000.
   Questo errore crashava il Worker nel catch path (finishTask
   lanciava dopo failTask), causando [Conductor] Worker crashed.

La combinazione dei due bug è stata la causa root dei retry a catena
sull'ordine VERALLI: il task completava su ERP+DB ma poi metrics.
finishTask lanciava → worker catch → failTask → retry → nuovo ordine
ERP (esacerbato dal bug db_committed→NULL già fixato in 5a34d158).

+++ archibald-web-app/backend/src/db/repositories/bot-metrics.ts
@@ -65,13 +65,13 @@ export async function recordTaskFinish(pool: Querier, params: TaskMetricFinish):
        error_message = $5,
        retry_count = $6,
        order_id = COALESCE($7, order_id),
-       ui_duration_ms = $9,
-       queue_wait_ms = EXTRACT(EPOCH FROM ($1 - enqueued_at)) * 1000,
-       bot_duration_ms = EXTRACT(EPOCH FROM ($2 - $1)) * 1000,
-       total_e2e_ms = $9 +
-                      EXTRACT(EPOCH FROM ($1 - enqueued_at)) * 1000 +
-                      EXTRACT(EPOCH FROM ($2 - $1)) * 1000
-     WHERE task_id = $8`,
+       ui_duration_ms = $9::bigint,
+       queue_wait_ms = EXTRACT(EPOCH FROM ($1 - enqueued_at))::bigint * 1000,
+       bot_duration_ms = EXTRACT(EPOCH FROM ($2 - $1))::bigint * 1000,
+       total_e2e_ms = $9::bigint +
+                      EXTRACT(EPOCH FROM ($1 - enqueued_at))::bigint * 1000 +
+                      EXTRACT(EPOCH FROM ($2 - $1))::bigint * 1000
+     WHERE task_id = $8::bigint`,
     [startedAt, completedAt, status, errorClass ?? null, errorMessage ?? null, retryCount, orderId ?? null, taskId.toString(), uiMs],
   );
 }
